+++ src/impls/boxed.rs
@@ -11,29 +11,31 @@ use crate::{DekuError, DekuReader, DekuWriter};
 impl<'a, T, Ctx> DekuReader<'a, Ctx> for Box<T>
 where
     T: DekuReader<'a, Ctx>,
-    Ctx: Copy,
+    Ctx: Clone,
 {
     fn from_reader_with_ctx<R: Read + Seek>(
         reader: &mut Reader<R>,
         inner_ctx: Ctx,
     ) -> Result<Self, DekuError> {
-        let val = <T>::from_reader_with_ctx(reader, inner_ctx)?;
+        let val = <T>::from_reader_with_ctx(reader, inner_ctx.clone())?;
         Ok(Box::new(val))
     }
 }
 
-impl<'a, T, Ctx, Predicate> DekuReader<'a, (Limit<T, Predicate>, Ctx)> for Box<[T]>
+impl<'a, T, Ctx, Predicate, PredicateWithCtx>
+    DekuReader<'a, (Limit<T, Predicate, Ctx, PredicateWithCtx>, Ctx)> for Box<[T]>
 where
     T: DekuReader<'a, Ctx>,
-    Ctx: Copy,
+    Ctx: Clone,
     Predicate: FnMut(&T) -> bool,
+    PredicateWithCtx: FnMut(&T, Ctx) -> bool,
 {
     fn from_reader_with_ctx<R: Read + Seek>(
         reader: &mut Reader<R>,
-        (limit, inner_ctx): (Limit<T, Predicate>, Ctx),
+        (limit, inner_ctx): (Limit<T, Predicate, Ctx, PredicateWithCtx>, Ctx),
     ) -> Result<Self, DekuError> {
         // use Vec<T>'s implementation and convert to Box<[T]>
-        let val = <Vec<T>>::from_reader_with_ctx(reader, (limit, inner_ctx))?;
+        let val = <Vec<T>>::from_reader_with_ctx(reader, (limit, inner_ctx.clone()))?;
         Ok(val.into_boxed_slice())
     }
 }
@@ -41,7 +43,7 @@ where
 impl<T, Ctx> DekuWriter<Ctx> for Box<[T]>
 where
     T: DekuWriter<Ctx>,
-    Ctx: Copy,
+    Ctx: Clone,
 {
     /// Write all `T`s to bits
     fn to_writer<W: Write + Seek>(
@@ -50,7 +52,7 @@ where
         ctx: Ctx,
     ) -> Result<(), DekuError> {
         for v in self.as_ref() {
-            v.to_writer(writer, ctx)?;
+            v.to_writer(writer, ctx.clone())?;
         }
         Ok(())
     }
@@ -59,7 +61,7 @@ where
 impl<T, Ctx> DekuWriter<Ctx> for Box<T>
 where
     T: DekuWriter<Ctx>,
-    Ctx: Copy,
+    Ctx: Clone,
 {
     /// Write all `T`s to bits
     fn to_writer<W: Write + Seek>(
@@ -67,7 +69,7 @@ where
         writer: &mut Writer<W>,
         ctx: Ctx,
     ) -> Result<(), DekuError> {
-        self.as_ref().to_writer(writer, ctx)?;
+        self.as_ref().to_writer(writer, ctx.clone())?;
         Ok(())
     }
 }
@@ -103,6 +105,9 @@ mod tests {
         assert_eq!(input.to_vec(), writer.inner.into_inner());
     }
 
+    type MyLimit<Predicate> =
+        Limit<u16, Predicate, (Endian, BitSize), fn(&u16, (Endian, BitSize)) -> bool>;
+
     // Note: Copied tests from vec.rs impl
     #[rstest(input, endian, bit_size, limit, expected, expected_rest_bits, expected_rest_bytes, expected_write,
         case::normal_le([0xAA, 0xBB, 0xCC, 0xDD].as_ref(), Endian::Little, Some(16), 2.into(), vec![0xBBAA, 0xDDCC].into_boxed_slice(), bits![u8, Msb0;], &[], vec![0xAA, 0xBB, 0xCC, 0xDD]),
@@ -116,7 +121,7 @@ mod tests {
         input: &[u8],
         endian: Endian,
         bit_size: Option<usize>,
-        limit: Limit<u16, Predicate>,
+        limit: MyLimit<Predicate>,
         expected: Box<[u16]>,
         expected_rest_bits: &bitvec::slice::BitSlice<u8, bitvec::prelude::Msb0>,
         expected_rest_bytes: &[u8],