+++ tests/test_attributes/test_ctx.rs
@@ -1,4 +1,5 @@
 use core::convert::{TryFrom, TryInto};
+use std::collections::{HashMap, HashSet};
 use std::io::Cursor;
 
 use deku::prelude::*;
@@ -285,4 +286,244 @@ fn test_enum_endian_ctx() {
 
     let ret_write: Vec<u8> = ret_read.try_into().unwrap();
     assert_eq!(ret_write, test_data)
+}
+
+#[test]
+fn test_interior_mutability_for_context_read_until_with_ctx() {
+    #[derive(Debug, Clone)]
+    struct IndexContext {
+        idx: std::rc::Rc<std::cell::Cell<usize>>,
+        n: usize,
+        fx: std::rc::Rc<std::cell::Cell<bool>>,
+    }
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone)]
+    struct A {
+        #[deku(
+            until_with_ctx = "|_:&B,ctx:IndexContext| !ctx.fx.get()",
+            ctx = "IndexContext { idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}",
+            writer_ctx = "IndexContext { idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: items.len(), fx: std::rc::Rc::new(std::cell::Cell::new(false)) }"
+        )]
+        items: Vec<B>,
+    }
+
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone)]
+    #[deku(
+        ctx = "ctx: IndexContext",
+        ctx_default = "IndexContext{idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}"
+    )] // this struct uses a context for serialization. For deserialization it also works with the default context.
+    struct B {
+        x: u8,
+        y: u8,
+        #[deku(
+            temp,
+            temp_value = "{let ret = ctx.idx.get() as u8; ctx.idx.set(ctx.idx.get()+1); ret}"
+        )]
+        idx_automatically_filled: u8,
+        #[deku(
+            read_post_processing = "ctx.fx.set(*auto_fx!=0);",
+            temp,
+            temp_value = "if ctx.idx.get() < ctx.n {1} else {0}"
+        )]
+        auto_fx: u8,
+    }
+
+    let test_data = A {
+        items: vec![B { x: 8, y: 9 }, B { x: 7, y: 9 }, B { x: 6, y: 9 }],
+    };
+
+    let ret_write: Vec<u8> = test_data.clone().try_into().unwrap();
+    assert_eq!(vec![8, 9, 0, 1, 7, 9, 1, 1, 6, 9, 2, 0], ret_write);
+    //                    ^  ^        ^  ^        ^  ^
+    //                    |  fx=1     |  fx=1     |  fx=0 (last)
+    //                   idx=0      idx=1       idx=2
+
+    // read the data back
+    let check_data = A::from_bytes((&ret_write, 0)).unwrap().1;
+    assert_eq!(check_data, test_data);
+
+    // check with fx=0 after the second element:
+    let check_data = A::from_bytes((&[8, 9, 0, 1, 7, 9, 1, 0], 0)).unwrap().1;
+    assert_eq!(check_data.items.len(), 2);
+}
+
+#[test]
+fn test_interior_mutability_for_context_read_until_with_ctx_hashset() {
+    #[derive(Debug, Clone)]
+    struct IndexContext {
+        idx: std::rc::Rc<std::cell::Cell<usize>>,
+        n: usize,
+        fx: std::rc::Rc<std::cell::Cell<bool>>,
+    }
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone)]
+    struct A {
+        #[deku(
+            until_with_ctx = "|_:&B,ctx:IndexContext| !ctx.fx.get()",
+            ctx = "IndexContext { idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}",
+            writer_ctx = "IndexContext { idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: items.len(), fx: std::rc::Rc::new(std::cell::Cell::new(false)) }"
+        )]
+        items: HashSet<B>,
+    }
+
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone, Eq, Hash)]
+    #[deku(
+        ctx = "ctx: IndexContext",
+        ctx_default = "IndexContext{idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}"
+    )] // this struct uses a context for serialization. For deserialization it also works with the default context.
+    struct B {
+        x: u8,
+        y: u8,
+        #[deku(
+            temp,
+            temp_value = "{let ret = ctx.idx.get() as u8; ctx.idx.set(ctx.idx.get()+1); ret}"
+        )]
+        idx_automatically_filled: u8,
+        #[deku(
+            read_post_processing = "ctx.fx.set(*auto_fx!=0);",
+            temp,
+            temp_value = "if ctx.idx.get() < ctx.n {1} else {0}"
+        )]
+        auto_fx: u8,
+    }
+
+    let test_data = A {
+        items: HashSet::from([B { x: 8, y: 9 }, B { x: 7, y: 9 }, B { x: 6, y: 9 }]),
+    };
+
+    let ret_write: Vec<u8> = test_data.clone().try_into().unwrap();
+    let test_bytes = vec![8, 9, 0, 1, 7, 9, 1, 1, 6, 9, 2, 0];
+    assert_eq!(ret_write.len(), test_bytes.len());
+    assert_eq!(ret_write.last().unwrap(), &0);
+
+    // read the data back
+    let check_data = A::from_bytes((&ret_write, 0)).unwrap().1;
+    assert_eq!(check_data, test_data);
+
+    // check with fx=0 after the second element:
+    let check_data = A::from_bytes((&[8, 9, 0, 1, 7, 9, 1, 0], 0)).unwrap().1;
+    assert_eq!(check_data.items.len(), 2);
+}
+
+#[test]
+fn test_interior_mutability_for_context_read_until_with_ctx_hashmap() {
+    #[derive(Debug, Clone)]
+    struct IndexContext {
+        idx: std::rc::Rc<std::cell::Cell<usize>>,
+        n: usize,
+        fx: std::rc::Rc<std::cell::Cell<bool>>,
+    }
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone)]
+    struct A {
+        #[deku(
+            until_with_ctx = "|x:&(N,B),ctx:IndexContext| !ctx.fx.get()",
+            ctx = "IndexContext { idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}",
+            writer_ctx = "IndexContext { idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: items.len(), fx: std::rc::Rc::new(std::cell::Cell::new(false)) }"
+        )]
+        items: HashMap<N, B>,
+    }
+
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone, Eq, Hash)]
+    #[deku(
+        ctx = "ctx: IndexContext",
+        ctx_default = "IndexContext{idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}"
+    )] // this struct uses a context for serialization. For deserialization it also works with the default context.
+    struct B {
+        x: u8,
+        y: u8,
+        #[deku(
+            temp,
+            temp_value = "{let ret = ctx.idx.get() as u8; ctx.idx.set(ctx.idx.get()+1); ret}"
+        )]
+        idx_automatically_filled: u8,
+        #[deku(
+            read_post_processing = "ctx.fx.set(*auto_fx!=0);",
+            temp,
+            temp_value = "if ctx.idx.get() < ctx.n {1} else {0}"
+        )]
+        auto_fx: u8,
+    }
+
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone, Eq, Hash)]
+    #[deku(
+        ctx = "ctx: IndexContext",
+        ctx_default = "IndexContext{idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}"
+    )] // this struct uses a context for serialization. For deserialization it also works with the default context.
+    struct N {
+        n: u8,
+    }
+
+    let test_data = A {
+        items: HashMap::from([
+            (N { n: 1 }, B { x: 8, y: 9 }),
+            (N { n: 2 }, B { x: 7, y: 9 }),
+            (N { n: 3 }, B { x: 6, y: 9 }),
+        ]),
+    };
+
+    let ret_write: Vec<u8> = test_data.clone().try_into().unwrap();
+    assert_eq!(ret_write.last().unwrap(), &0);
+
+    // read the data back
+    let check_data = A::from_bytes((&ret_write, 0)).unwrap().1;
+    assert_eq!(check_data, test_data);
+}
+
+#[test]
+fn test_interior_mutability_for_context_read_until_with_ctx_slice_ref() {
+    #[derive(Debug, Clone)]
+    struct IndexContext {
+        idx: std::rc::Rc<std::cell::Cell<usize>>,
+        n: usize,
+        fx: std::rc::Rc<std::cell::Cell<bool>>,
+    }
+
+    #[deku_derive(DekuRead, DekuWrite)]
+    #[derive(PartialEq, Debug, Clone)]
+    #[deku(
+        ctx = "ctx: IndexContext",
+        ctx_default = "IndexContext{idx: std::rc::Rc::new(std::cell::Cell::new(0)), n: 0, fx: std::rc::Rc::new(std::cell::Cell::new(false))}"
+    )] // this struct uses a context for serialization. For deserialization it also works with the default context.
+    struct B {
+        x: u8,
+        y: u8,
+        #[deku(
+            temp,
+            temp_value = "{let ret = ctx.idx.get() as u8; ctx.idx.set(ctx.idx.get()+1); ret}"
+        )]
+        idx_automatically_filled: u8,
+        #[deku(
+            read_post_processing = "ctx.fx.set(*auto_fx!=0);",
+            temp,
+            temp_value = "if ctx.idx.get() < ctx.n {1} else {0}"
+        )]
+        auto_fx: u8,
+    }
+
+    let data = [B { x: 8, y: 9 }, B { x: 7, y: 9 }, B { x: 6, y: 9 }];
+
+    let mut buf = vec![];
+    let mut cursor = Cursor::new(&mut buf);
+    let mut writer = Writer::new(&mut cursor);
+    let myref = &data[..];
+    // auto referencing rules: we want to call the method on `&[T]`:
+    (&myref)
+        .to_writer(
+            &mut writer,
+            IndexContext {
+                idx: std::rc::Rc::new(std::cell::Cell::new(0)),
+                n: myref.len(),
+                fx: std::rc::Rc::new(std::cell::Cell::new(false)),
+            },
+        )
+        .unwrap();
+    assert_eq!(vec![8, 9, 0, 1, 7, 9, 1, 1, 6, 9, 2, 0], buf);
+    //                    ^  ^        ^  ^        ^  ^
+    //                    |  fx=1     |  fx=1     |  fx=0 (last)
+    //                   idx=0      idx=1       idx=2
 }